Breakout api.py into a package
